Add files via upload
Project - 2

The core idea behind this project is to represent an image as a one-dimensional vector, which can be thought of as a long sequence of numbers. When images are damaged or compressed, valuable information may be lost, making the image appear distorted or incomplete. By applying sophisticated algorithms and mathematical transformations to these one-dimensional vectors, the project attempts to recover and restore the original image's quality and content.